+++ src/types/index.ts
@@ -1,2 +1,3 @@
-export { LogLevel } from './LogLevel'
-export { PayloadType } from './PayloadType'
+export * from './LogLevel'
+export * from './PayloadParameters'
+export * from './PayloadType'