+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- The `Payload` class is now part of the countertop.  This will replace the `@tvkitchen/base-classes` version of Payload.
+- TypeScript definitions related to `Payload`, `PayloadType`, and `PayloadParameters` are also exported.
 
 ## [0.4.1] - 2022-06-08
 ### Changed